why sample size is treated as a long in esdlib, but an int everywhere else
	* esd.h, esdlib.c: why sample size is treated as a long in esdlib,
 	but an int everywhere else is beyond me.  64 bit platforms should
 	be happier now. still have to check 32-64 bit interoperability

	* configure.in: touch-ups and commenting *shudder*.

+++ esd.h
@@ -121,7 +121,7 @@ int esd_record_stream( esd_format_t format, int rate, char *host, char *name );
 int esd_record_stream_fallback( esd_format_t format, int rate, char *host, char *name );
 
 /* cache a sample in the server returns sample id, < 0 = error */
-int esd_sample_cache( int esd, esd_format_t format, int rate, long length, char *name );
+int esd_sample_cache( int esd, esd_format_t format, int rate, int length, char *name );
 int esd_confirm_sample_cache( int esd );
 
 /* get the sample id for an already-cached sample */
